requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
pandas==2.1.4
google-generativeai==0.3.2
python-dotenv==1.0.0